requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# AWS
boto3>=1.34.0

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0
alembic>=1.13.0
aiosqlite>=0.19.0

# Date Parsing
python-dateutil>=2.8.2

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0
